refactor: rename manager
